+++ hw01/src/main/java/ru/otus/spring/dao/QuestionsDao.java
@@ -3,8 +3,10 @@ package ru.otus.spring.dao;
 import java.util.List;
 import java.util.Map;
 
+@SuppressWarnings("checkstyle:LeftCurly")
 public interface QuestionsDao
 {
     Map<String,List<String>> getAll();
+
     List<String> findByQuestions(String questionName);
 }
