+++ hw01/src/test/resources/questions_test.csv
@@ -1,0 +1,3 @@
+question1,q1_text1
+question2,q2_text1,q2_text2
+question3,
